Add files via upload
Introducing the CodeCraft Calculator: Your gateway to seamless HTML mastery! Crafted with care for beginners, this tool is not just about numbers; it's about unlocking the magic of coding. Dive into the world of HTML with this sleek calculator, designed to make learning an exhilarating journey. 
Let the keystrokes pave the way to your coding adventure!